requirements.txt
beautifulsoup4==4.12.3
bs4~=0.0.2
Requests==2.31.0
matplotlib~=3.8.2
nltk~=3.8.1
Unidecode==1.3.8
wordcloud~=1.9.3

pillow~=10.2.0
